+++ helper/src/main/java/com/vsg/helper/common/util/viewModel/viewRoom/IViewModelAllPagingViewRoom.kt
@@ -6,5 +6,5 @@ import kotlinx.coroutines.flow.Flow
 
 interface IViewModelAllPagingViewRoom<TEntity>
         where TEntity : IEntityViewRoom<TEntity> {
-    fun viewModelGetViewAllPaging(idRelation: Long): Flow<PagingData<TEntity>>
+    fun viewModelGetViewAllPaging(idRelation: Int): Flow<PagingData<TEntity>>
 }
